SystemCode/backend/requirements.txt
# web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1

# pydantic core & settings
pydantic==2.6.4
pydantic-settings==2.2.1

# database
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# env config
python-dotenv==1.0.0

# cache
redis==5.0.1

# llm client
openai==1.12.0
httpx==0.26.0